app/requirements.txt
altgraph==0.10.2
appnope==0.1.0
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.9.3
Automat==0.6.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
bdist-mpkg==0.5.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bonjour-py==0.3
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
decorator==4.1.2
Django==1.11.6
django-mako==0.1.5rc0
docutils==0.14
enum34==1.1.6
Fabric==1.14.0
Flask==0.12.2
Flask-Mako==0.4
Flask-SQLAlchemy==2.3.2
future==0.16.0
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.19
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
lockfile==0.12.2
lxml==4.1.1
macholib==1.5.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==1.3.1
meld3==1.0.2
modulegraph==0.10.4
mysql-connector-python-cext==8.0.5
numpy==1.8.0rc1
paramiko==2.4.0
parsel==1.3.1
pathlib2==2.3.0
pbr==3.1.1
pep8==1.7.1
pexpect==4.3.0
pickleshare==0.7.4
prompt==0.4.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py2app==0.7.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyJWT==1.5.3
PyNaCl==1.2.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==17.5.0
pyparsing==2.0.1
python-daemon==2.1.2
python-dateutil==1.5
pytz==2013.7
queuelib==1.4.2
requests==2.18.4
scandir==1.6
scipy==0.13.0b1
Scrapy==1.5.0
service==0.4.1
service-identity==17.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.2
sqlparse==0.2.4
stevedore==1.27.1
supervisor==3.3.3
Tempita==0.5.2
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
urllib3==1.22
uWSGI==2.0.15
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
w3lib==1.18.0
wcwidth==0.1.7
Werkzeug==0.14.1
xattr==0.6.4
zope.interface==4.4.3
